Move some helper methods around
