+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/effects/debugger_effects.ts
@@ -94,7 +94,10 @@ import {
   getPollSilenceTimeMs,
   getSourceFileListLoaded,
 } from '../store/debugger_selectors';
-import {beginEndRangesInclude} from '../store/debugger_store_utils';
+import {
+  beginEndRangesInclude,
+  stackFrameAsArray2StackFrame,
+} from '../store/debugger_store_utils';
 import {
   DataLoadState,
   DebuggerRunListing,
@@ -638,18 +641,9 @@ export class DebuggerEffects {
             // TODO(cais): Do this reshaping in the backend and simplify
             // the frontend code here.
             for (let i = 0; i < stackFrameIds.length; ++i) {
-              const [
-                host_name,
-                file_path,
-                lineno,
-                function_name,
-              ] = stackFramesResponse.stack_frames[i];
-              stackFramesById[stackFrameIds[i]] = {
-                host_name,
-                file_path,
-                lineno,
-                function_name,
-              };
+              stackFramesById[stackFrameIds[i]] = stackFrameAsArray2StackFrame(
+                stackFramesResponse.stack_frames[i]
+              );
               stackFramesResponse.stack_frames[i];
             }
             this.store.dispatch(
@@ -848,18 +842,11 @@ export class DebuggerEffects {
               // TODO(cais): Do this reshaping in the backend and simplify
               // the frontend code here.
               for (let i = 0; i < missingStackFrameIds.length; ++i) {
-                const [
-                  host_name,
-                  file_path,
-                  lineno,
-                  function_name,
-                ] = stackFramesResponse.stack_frames[i];
-                stackFramesById[missingStackFrameIds[i]] = {
-                  host_name,
-                  file_path,
-                  lineno,
-                  function_name,
-                };
+                stackFramesById[
+                  missingStackFrameIds[i]
+                ] = stackFrameAsArray2StackFrame(
+                  stackFramesResponse.stack_frames[i]
+                );
               }
               this.store.dispatch(
                 stackFramesLoaded({stackFrames: stackFramesById})
